make utils an internal module and add yaml config file parser
